refactor: update main
