--- src/main/factories/authentication/index.ts
@@ -1,1 +1,0 @@
-export * from '@/main/factories/authentication/authentication-factory';
