--- src/main/factories/delete-user-by-id/index.ts
@@ -1,1 +1,0 @@
-export * from '@/main/factories/delete-user-by-id/delete-user-by-id-factory';
