--- src/main/factories/delete-user/delete-user-factory.ts
@@ -1,11 +1,0 @@
-import { DeleteUserUseCase } from '@/use-cases/delete-user';
-import { DeleteUserController } from '@/adapters/controllers/delete-user-controller';
-import { IController } from '@/adapters/interfaces';
-import { UserRepository } from '@/infra/repositories';
-
-export function makeDeleteUserController(): IController {
-  const userRepository = new UserRepository();
-  const deleteUserUseCase = new DeleteUserUseCase(userRepository);
-  const deleteUserController = new DeleteUserController(deleteUserUseCase);
-  return deleteUserController;
-}
