--- src/main/factories/delete-user/index.ts
@@ -1,1 +1,0 @@
-export * from '@/main/factories/delete-user/delete-user-factory';
