--- src/main/factories/external-sign-in/external-sign-in-validation-factory.ts
@@ -1,3 +1,0 @@
-import { ValidationComposite, RequiredFieldsValidation } from '@/adapters/util/validations';
-
-export const makeExternalSignInValidationFactory = (): ValidationComposite => new ValidationComposite([new RequiredFieldsValidation(['name', 'email'])]);
