--- src/main/factories/external-sign-in/index.ts
@@ -1,2 +1,0 @@
-export * from '@/main/factories/external-sign-in/external-sign-in-validation-factory';
-export * from '@/main/factories/external-sign-in/external-sign-in-factory';
