--- src/main/factories/read-all-users/index.ts
@@ -1,1 +1,0 @@
-export * from '@/main/factories/read-all-users/read-all-users-factory';
