--- src/main/factories/read-user-by-id/index.ts
@@ -1,1 +1,0 @@
-export * from '@/main/factories/read-user-by-id/read-user-by-id-factory';
