--- src/main/factories/read-user/index.ts
@@ -1,1 +1,0 @@
-export * from '@/main/factories/read-user/read-user-factory';
