--- src/main/factories/sign-in/index.ts
@@ -1,2 +1,0 @@
-export * from '@/main/factories/sign-in/sign-in-validation-factory';
-export * from '@/main/factories/sign-in/sign-in-factory';
