--- src/main/factories/sign-in/sign-in-factory.ts
@@ -1,17 +1,0 @@
-import { SignInUseCase } from '@/use-cases/sign-in';
-import { SignInController } from '@/adapters/controllers/sign-in-controller';
-import { IController } from '@/adapters/interfaces';
-import { UserRepository } from '@/infra/repositories';
-import { BcryptAdapter, JwtAdapter } from '@/infra/criptography';
-import { makeSignInValidationFactory } from '@/main/factories/sign-in';
-import env from '@/main/config/env';
-
-export function makeSignInController(): IController {
-  const salt = env.bcryptSalt;
-  const userRepository = new UserRepository();
-  const bcryptAdapter = new BcryptAdapter(salt);
-  const jwtAdapter = new JwtAdapter(env.jwtSecret, env.expiresIn);
-  const signInUseCase = new SignInUseCase(userRepository, bcryptAdapter, jwtAdapter);
-  const signInController = new SignInController(makeSignInValidationFactory(), signInUseCase);
-  return signInController;
-}
