--- src/main/factories/sign-in/sign-in-validation-factory.ts
@@ -1,5 +1,0 @@
-import { ValidationComposite, RequiredFieldsValidation } from '@/adapters/util/validations';
-
-export const makeSignInValidationFactory = (): ValidationComposite => new ValidationComposite([
-  new RequiredFieldsValidation(['email', 'password']),
-]);
