--- src/main/factories/sign-up/index.ts
@@ -1,2 +1,0 @@
-export * from '@/main/factories/sign-up/sign-up-validation-factory';
-export * from '@/main/factories/sign-up/sign-up-factory';
