--- src/main/factories/update-user-by-id/index.ts
@@ -1,1 +1,0 @@
-export * from '@/main/factories/update-user-by-id/update-user-by-id-factory';
