--- src/main/factories/update-user/index.ts
@@ -1,1 +1,0 @@
-export * from '@/main/factories/update-user/update-user-factory';
