+++ src/main/factories/user/authentication/index.ts
@@ -1,0 +1,1 @@
+export * from '@/main/factories/user/authentication/authentication-factory';
