+++ src/main/factories/user/delete-user-by-id/index.ts
@@ -1,0 +1,1 @@
+export * from '@/main/factories/user/delete-user-by-id/delete-user-by-id-factory';
