+++ src/main/factories/user/delete-user/index.ts
@@ -1,0 +1,1 @@
+export * from '@/main/factories/user/delete-user/delete-user-factory';
