+++ src/main/factories/user/external-sign-in/external-sign-in-validation-factory.ts
@@ -1,0 +1,3 @@
+import { ValidationComposite, RequiredFieldsValidation } from '@/adapters/controllers/util/validations';
+
+export const makeExternalSignInValidationFactory = (): ValidationComposite => new ValidationComposite([new RequiredFieldsValidation(['name', 'email'])]);
