+++ src/main/factories/user/external-sign-in/index.ts
@@ -1,0 +1,2 @@
+export * from '@/main/factories/user/external-sign-in/external-sign-in-validation-factory';
+export * from '@/main/factories/user/external-sign-in/external-sign-in-factory';
