+++ src/main/factories/user/index.ts
@@ -1,0 +1,11 @@
+export * from '@/main/factories/user/sign-up';
+export * from '@/main/factories/user/sign-in';
+export * from '@/main/factories/user/external-sign-in';
+export * from '@/main/factories/user/read-user';
+export * from '@/main/factories/user/read-all-users';
+export * from '@/main/factories/user/read-user-by-id';
+export * from '@/main/factories/user/update-user';
+export * from '@/main/factories/user/update-user-by-id';
+export * from '@/main/factories/user/delete-user';
+export * from '@/main/factories/user/delete-user-by-id';
+export * from '@/main/factories/user/authentication';
