+++ src/main/factories/user/read-all-users/index.ts
@@ -1,0 +1,1 @@
+export * from '@/main/factories/user/read-all-users/read-all-users-factory';
