+++ src/main/factories/user/read-user-by-id/index.ts
@@ -1,0 +1,1 @@
+export * from '@/main/factories/user/read-user-by-id/read-user-by-id-factory';
