+++ src/main/factories/user/read-user/index.ts
@@ -1,0 +1,1 @@
+export * from '@/main/factories/user/read-user/read-user-factory';
