+++ src/main/factories/user/sign-in/sign-in-validation-factory.ts
@@ -1,0 +1,5 @@
+import { ValidationComposite, RequiredFieldsValidation } from '@/adapters/controllers/util/validations';
+
+export const makeSignInValidationFactory = (): ValidationComposite => new ValidationComposite([
+  new RequiredFieldsValidation(['email', 'password']),
+]);
