+++ src/main/factories/user/sign-up/index.ts
@@ -1,0 +1,2 @@
+export * from '@/main/factories/user/sign-up/sign-up-validation-factory';
+export * from '@/main/factories/user/sign-up/sign-up-factory';
