+++ src/main/factories/user/sign-up/sign-up-validation-factory.ts
@@ -1,0 +1,6 @@
+import { ValidationComposite, RequiredFieldsValidation, CompareFieldsValidation } from '@/adapters/controllers/util/validations';
+
+export const makeSignUpValidationFactory = (): ValidationComposite => new ValidationComposite([
+  new RequiredFieldsValidation(['name', 'email', 'password', 'passwordConfirmation']),
+  new CompareFieldsValidation('password', 'passwordConfirmation'),
+]);
