+++ src/main/factories/user/update-user-by-id/index.ts
@@ -1,0 +1,1 @@
+export * from '@/main/factories/user/update-user-by-id/update-user-by-id-factory';
