+++ src/main/factories/user/update-user/index.ts
@@ -1,0 +1,1 @@
+export * from '@/main/factories/user/update-user/update-user-factory';
