--- src/main/middlewares/authentication.ts
@@ -1,4 +1,0 @@
-import { middlewareAdapter } from '@/main/adapters/express';
-import { makeAuthenticationMiddleware } from '@/main/factories/authentication';
-
-export const authentication = middlewareAdapter(makeAuthenticationMiddleware());
