+++ src/main/middlewares/index.ts
@@ -1,5 +1,3 @@
 export * from '@/main/middlewares/body-parser';
 export * from '@/main/middlewares/cors';
 export * from '@/main/middlewares/content-type';
-export * from '@/main/middlewares/authentication';
-export * from '@/main/middlewares/googleOAuth';
