+++ src/main/middlewares/user/authentication.ts
@@ -1,0 +1,4 @@
+import { middlewareAdapter } from '@/main/adapters/express';
+import { makeAuthenticationMiddleware } from '@/main/factories/user';
+
+export const authentication = middlewareAdapter(makeAuthenticationMiddleware());
